Rename and move common functions in loaders and tests
